staking(validator_manager): scaffolding for `ValidatorDataRead`
