+++ crates/core/component/stake/src/component/action_handler/undelegate.rs
@@ -6,8 +6,8 @@ use cnidarium::{StateRead, StateWrite};
 use penumbra_shielded_pool::component::SupplyWrite;
 
 use crate::{
-    component::action_handler::ActionHandler, component::StateWriteExt as _, event, StateReadExt as _,
-    Undelegate,
+    component::action_handler::ActionHandler, component::StateWriteExt as _, event,
+    StateReadExt as _, Undelegate,
 };
 
 #[async_trait]
