+++ crates/core/component/stake/src/component/rpc.rs
@@ -15,7 +15,7 @@ use penumbra_proto::{
 use tonic::Status;
 use tracing::instrument;
 
-use super::StateReadExt;
+use super::{validator_manager::ValidatorDataRead, StateReadExt};
 use crate::validator;
 
 // TODO: Hide this and only expose a Router?
